Chef 14 compatibility fixes (#9)
* Remove deprecated coverage reports

* Change versioning to major only

The newer version of ChefSpec and Fauxhai we have installed will not
automatically choose the newest minor version for you.

* Change log level from :fatal to :warn

This is to ensure we see all compatibility warnings. In the past this was a temporary change. However, moving forward we want to keep the warning enabled so that we can deal with deprecation issues sooner.

* Remove deprecated Style/IfUnlessModifier from rubocop config file